requirements.txt
streamlit
ccxt
pandas
numpy
python-dotenv
requests
